+++ codeclimate/codeclimate.go
@@ -1,4 +1,4 @@
-package main
+package codeclimate
 
 import (
 	"encoding/json"
