+++ main.go
@@ -1,6 +1,8 @@
 package main
 
 import (
+	"github.com/eorate/shadowheart/codeclimate"
+
 	"github.com/gin-gonic/gin"
 )
 
@@ -8,7 +10,7 @@ func setupRouter() *gin.Engine {
 	r := gin.Default()
 
 	r.GET("/metrics", func(c *gin.Context) {
-		c.JSON(200, BuildRepositoryStats())
+		c.JSON(200, codeclimate.BuildRepositoryStats())
 	})
 	return r
 }
