+++ mocks/mocks.go
@@ -1,4 +1,4 @@
-package main
+package mocks
 
 import "net/http"
